requirements.txt
# Core ML
ultralytics==8.3.40
torch==2.5.1
torchvision==0.20.1

# Image processing
opencv-python==4.10.0.84
Pillow==11.0.0
numpy==1.26.4
albumentations==1.4.21

# Dataset scraping
duckduckgo-search==6.3.7
requests==2.32.3

# API serving
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
slowapi==0.1.9

# Config / validation
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
PyYAML==6.0.2

# Demo
gradio==5.9.1

# Utilities
tqdm==4.67.1
pandas==2.2.3
matplotlib==3.9.3
scipy==1.14.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Security audit (run before deploy)
pip-audit==2.7.3
